make addCards testable
